requirements.txt
streamlit
groq
sentence-transformers
chromadb
SPARQLWrapper
pandas
numpy
kagglehub
python-dotenv
